requirements.txt
paho-mqtt
requests
python-dotenv
Flask
plotly 